+++ packages/lambda-tiler/src/tile.set.cache.ts
@@ -67,7 +67,7 @@ export class TileSetCache {
             return existing;
         }
 
-        if (Config.TileSet.isRaster(tileSet)) {
+        if (Config.isTileSetRaster(tileSet)) {
             const ts = new TileSetRaster(name, tileMatrix);
             await ts.init(tileSet);
             this.tileSets.set(tileSetId, ts);
@@ -86,8 +86,8 @@ export class TileSetCache {
 
         const promises: Promise<TileSet | null>[] = [];
         for (const tileMatrix of tileMatrices) promises.push(this.get(name, tileMatrix));
-
         const tileMatrixSets = await Promise.all(promises);
+
         const tileSets: TileSetRaster[] = [];
         for (const parent of tileMatrixSets) {
             if (parent == null) continue;